+++ compression.h
@@ -16,12 +16,12 @@ class Compressor {
     Schema schema_;
     std::unique_ptr<ModelLearner> learner_;
     std::vector< std::unique_ptr<Model> > model_;
-    std::vector<int> attr_order_;
+    std::vector<size_t> attr_order_;
     std::unique_ptr<ByteWriter> byte_writer_;
     int stage_;
-    int num_of_tuples_;
-    int implicit_prefix_length_;
-    std::vector<int> block_length_;
+    size_t num_of_tuples_;
+    size_t implicit_prefix_length_;
+    std::vector<size_t> block_length_;
   public:
     // sort_by_attr_index_ equals -1 meaning no specific sorting required
     Compressor(char* outputFile, const Schema& schema, const CompressionConfig& config);
