+++ data_io.cpp
@@ -46,18 +46,18 @@ TupleOStream& operator>>(TupleOStream& tuple_stream, int& val) {
  * Since the prefix/suffix has at most 7 bits, we can use 0xff as special symbol to indicate
  * that the prefix/suffix is not used yet.
  */
-ByteWriter::ByteWriter(std::vector<int>* block_length, const std::string& file_name) :
+ByteWriter::ByteWriter(std::vector<size_t>* block_length, const std::string& file_name) :
     block_unwritten_prefix_(block_length->size()),
     block_unwritten_suffix_(block_length->size(), 0xff),
     file_(file_name) {
     block_pos_.swap(*block_length);
-    int total_len = 0;
-    for (int i = 0; i < block_pos_.size(); i++) {
-        int current_block_len = block_pos_[i];
+    size_t total_len = 0;
+    for (size_t i = 0; i < block_pos_.size(); i++) {
+        size_t current_block_len = block_pos_[i];
         block_pos_[i] = total_len;
         total_len += current_block_len;
     }
-    for (int i = 0; i < block_pos_.size(); i++) {
+    for (size_t i = 0; i < block_pos_.size(); i++) {
         if ((block_pos_[i] & 7) == 0)
             block_unwritten_suffix_[i] = 0;
     }
@@ -66,11 +66,11 @@ ByteWriter::ByteWriter(std::vector<int>* block_length, const std::string& file_n
 // Write all the remaining, we can use block_pos_ to identify the prefix/suffix length
 // of corresponding blocks
 ByteWriter::~ByteWriter() {
-    for (int i = 1; i < block_pos_.size(); i++ )
+    for (size_t i = 1; i < block_pos_.size(); i++ )
     if ((block_pos_[i - 1] & 7) != 0) {
-        int suffix_length = (block_pos_[i - 1] & 7);
+        size_t suffix_length = (block_pos_[i - 1] & 7);
         if (block_pos_[i] < block_pos_[i - 1] + (8 - suffix_length)) {
-            int len = block_pos_[i] - block_pos_[i - 1];
+            size_t len = block_pos_[i] - block_pos_[i - 1];
             block_unwritten_suffix_[i] = 
                 (block_unwritten_suffix_[i - 1] << len)
                 | (block_unwritten_prefix_[i]);
@@ -84,16 +84,16 @@ ByteWriter::~ByteWriter() {
     }
 }
 
-void ByteWriter::WriteByte(unsigned char byte, int block) {
+void ByteWriter::WriteByte(unsigned char byte, size_t block) {
     WriteLess(byte, 8, block);
 }
 
-void ByteWriter::WriteLess(unsigned char byte, int len, int block) {
+void ByteWriter::WriteLess(unsigned char byte, size_t len, size_t block) {
     file_.seekp(block_pos_[block] >> 3, std::ios_base::beg);
     byte = (byte & ((1 << len) - 1)); 
     if (block_unwritten_suffix_[block] != 0xff) {
         unsigned char suffix = block_unwritten_suffix_[block];
-        int needed_len = 8 - (block_pos_[block] & 7);
+        size_t needed_len = 8 - (block_pos_[block] & 7);
         if (needed_len <= len) {
             len -= needed_len;
             file_.put((suffix << needed_len) | (byte >> len));
@@ -106,7 +106,7 @@ void ByteWriter::WriteLess(unsigned char byte, int len, int block) {
         }
     }
     else {
-        int needed_len = 8 - (block_pos_[block] & 7);
+        size_t needed_len = 8 - (block_pos_[block] & 7);
         if (needed_len <= len) {
             len -= needed_len;
             unsigned char pad = (byte >> len);
