+++ data_io.h
@@ -61,14 +61,14 @@ class ByteWriter {
     std::vector<unsigned char> block_unwritten_prefix_;
     std::vector<unsigned char> block_unwritten_suffix_;
     // The starting point of incoming bit stream for each block
-    std::vector<int> block_pos_;
+    std::vector<size_t> block_pos_;
     std::ofstream file_;
   public:
-    ByteWriter(std::vector<int>* block_length, const std::string& file_name);
+    ByteWriter(std::vector<size_t>* block_length, const std::string& file_name);
     ~ByteWriter();
-    void WriteByte(unsigned char byte, int block);
+    void WriteByte(unsigned char byte, size_t block);
     // Only write the least significant (len) bits
-    void WriteLess(unsigned char byte, int len, int block);
+    void WriteLess(unsigned char byte, size_t len, size_t block);
 };
 
 }  // namespace db_compress
